feat: Update Justfile to change default serial port to /dev/a1 and add print command for joint angles
docs: Add README for DataCoach setup and usage instructions

feat: Add ZMQ_POLICY_ACTION_PORT constant for policy action output

feat: Introduce ReplayAction enum in data collector for replay actions

feat: Enhance DataCollector with abort handling and user prompts for replay actions

fix: Change action data type in ZMQPolicyServer to float32 for compatibility

docs: Update SETUP_UDEV.md for stable /dev/a1 symlink creation

fix: Modify dragdatacoach.sh to default to /dev/a1 for driver launch

feat: Improve dragdatacoach_all_in_one.sh with enhanced error handling and retry logic

feat: Add script to print joint angles from ROS topic

feat: Update run_data_collection.py to support multiple collection sessions with user prompts

feat: Refactor run_drag_replay_collection.py for improved error handling and replay session management

feat: Create doctor.py script to check required modules and environment setup for inference

fix: Update my_serve_policy.py to handle optional websocket inference backend

fix: Adjust serve_policy.py to conditionally import websocket_policy_server

+++ datacoach/inference/ZMQ_policy_server.py
@@ -1,3 +1,4 @@
+import cv2
 import zmq
 import numpy as np
 import time
@@ -113,7 +114,7 @@ class ZMQPolicyServer:
         obs = {
             "cam_0": image,
             "state": state,
-            "action": np.zeros(state.shape, dtype=np.uint8),
+            "action": np.zeros(state.shape, dtype=np.float32),
             "prompt": "do something",
         }
